pyproject.toml
[tool.poetry]
name = "mieet-app"
version = "0.1.1"
description = ""
authors = ["Marco Silva <mail@marco-silva.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
python-telegram-bot = "^12.0.0"
python-dotenv = "^0.10.3"
psycopg2 = "^2.8.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
